+++ .samignore
@@ -1,0 +1,4 @@
+# Ignore editor temp files
+*.pyc
+*.pyo
+.#*
